refactor(executor): split executor.go into executor.go + dispatch.go
executor.go (180 lines) retains types, sentinel errors, Option funcs,
GRPCExecutor struct, and NewGRPCExecutor constructor.

dispatch.go (388 lines) holds the execution logic: ExecuteTask,
resolveDeadline, dispatch, parseTokensUsed, isTransient, and Close.

Both files are under the 500-line threshold. Remove
internal/executor/executor.go from the grandfathered exception list in
scripts/checks/file_size.py — the file_size CI guard now covers the
executor package without exceptions.
