+++ FILEMAP.md
@@ -1,14 +1,14 @@
 # Repository File Map
 
 > Auto-generated by `scripts/generate_filemap.py` — 2026-04-20  
-> **270** tracked files
+> **271** tracked files
 
 ## Summary
 
 | Top-level directory | Files |
 |---------------------|-------|
 | `docs/` | 55 |
-| `internal/` | 52 |
+| `internal/` | 53 |
 | `agents/` | 45 |
 | `tests/` | 35 |
 | `(root)` | 20 |
@@ -28,7 +28,7 @@
 |-----------|-------|
 | `.py` | 91 |
 | `.md` | 71 |
-| `.go` | 48 |
+| `.go` | 49 |
 | `.yaml` | 23 |
 | `.rs` | 7 |
 | `.yml` | 6 |
@@ -254,6 +254,7 @@
 │   │   ├── defaults.go
 │   │   └── defaults_test.go
 │   ├── executor/
+│   │   ├── dispatch.go
 │   │   ├── executor.go
 │   │   └── executor_test.go
 │   ├── generated/
